// good-haul-back-end/src/auth/dto/user.login.dto.ts
/*
 * @FilePath: login.dto.ts
 * @Author: Aron
 * @Date: 2024-03-24 16:48:27
 * @LastEditors:
 * @LastEditTime: 2024-03-24 16:48:28
 * Copyright: 2024 xxxTech CO.,LTD. All Rights Reserved.
 * @Descripttion:
 */
// login.dto.ts
export class UserPasswordLoginDto {
  phone: string;
  password: string;
}

export class UserCodeLoginDto {
  phone: string;
  code: string;
}